requirements.txt
# Core LangChain and Google GenAI dependencies
langchain>=0.3.0
langchain-community>=0.0.148
langchain-google-genai>=0.1.6
google-generativeai>=0.8.5
langgraph>=0.2.3
langchain-experimental>=0.0.148

# Web and App Frameworks
streamlit>=1.32.0
streamlit-chat>=0.0.4
gradio>=5.9.1
fastapi
uvicorn

# Utilities & Environment
python-dotenv>=1.0.1
requests>=2.31.0
faker

# NLP / Embeddings / Vector Search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# PDF / Document Handling
PyPDF2>=3.0.1

#Database & SQL Utilities
sqlalchemy
sqlparse
pydantic

# Other
ddgs # DuckDuckGo Search package
tiktoken
bs4
IPython